Implement parsing MultiProcessor tables, initializing APIC, initializing SMP, and port over the interrupt handling in Rust from eduOS-rs.
MAX_CORES is reduced to 256, because MultiProcessor tables don't support more than that.
entry.asm is relicensed to MIT just like the rest of the eduOS-rs and HermitCore Rust code.
Also improve debugging and make use of the or_else/unwrap_or_else Rust functions.

This is all untested work in progress! At least it compiles :)