Commit 5e3ed9f76535a498e8ef37d8e1040c2311daa7f9 for the DoublyLinkedList implemented in Safe Rust.
Note that the unsafe implementation uses pointers and could call internal_allocate to allocate a new
node without locking. On the other hand, the safe implementation uses Rc, which always calls into our
allocator the same way.

Therefore, the global MM_LOCK was replaced by a custom MmLock that checks if a call originates from the same
CPU core. In that case, no further locking is done. This prevents deadlocks when doing recursive calls
from the Memory Manager into other Memory Manager functions.

Recursive calls also need to be handled in the NodePool.
When a maintenance is already in progress, a recursive call into maintain() must exit immediately.

With these changes, we now have a stable Memory Manager implemented in two versions:
Safe with Rust allocations and borrow-checking as well as unsafe with traditional pointers.
They can now be compared in a further step.