Implement ACPI support that parses the MADT for Processor/Local APIC information and the FADT/DSDT/SSDT for ACPI Power Off information.
This supersedes getting processor information from the MultiProcessor Specification 1.4 tables.
The advantage are broader hardware support (not all systems are MPS1.4-compliant) and support for logical processors (SMT/HyperThreading), which are only exposed in the MADT.

ACPI Power Off also offers a universal way to shut down a computer, no matter if it's a real one or a virtual machine.
It will be used in the following commits.

So far, this has only been tested under QEMU. Real hardware testing follows next.