merge all repositories to one
Internally we had many discussions about repositories / branch names
(see for instance hermitcore/libhermit-rs#35). This PR tries to merge
all different repositories/branches to one. If we do this, we have
to archive the RustyHermit repo and link all crates to this repo.
In addition, we have to revise our C examples in HermitPlaygroud.