Introduce integration tests
Adds an initial version of integration tests. The tests are located in
the top level of the 'tests' directory. Integration tests are executed via
the hermit_test_runner.py script, which can use either uhyve or qemu to execute
the tests.
In this initial version, using the test_harness is deactivated.