Changed endian types! Now carry native unsigned.
The little and big endian types be.. and le.. for u16, u32 and u64 now
carry a natively endian encoded unsigned integer. This allows to easily
compute mathematical operations with theses types and reduces the amount
of otherwise needed byte-swappings.

In order to get a little or big respectively endian type one must the
provide ::as_u{16, 32, 64} function.

The types now only indicate which endian value is expected by the
standard and by the hardware and probably only needs to be converted
into a real little or big endian type upon communication with the pci
configuration space!