Finished mappings for Common cfg & notif cfg
The ComCfgRaw struct is mapped and wrapped inside a box. The Box
is held by the public wrapper ComCfg struct.

The NotifCfg struct has a different layout, directly inside the
configuration space (NOT the memory are of the device). This made some
changes on the PciCap struct necessary.
The PciCap now carries a Origin structure which will allow drivers and
other structures to gain access to the original PCI capability inside
the devices configuration space.
The NotifCfg struct now allows to write to a single byte inside the
configuration space. The byte hereby represents the INT#X flags.