Map shared memory, fixes wrong mappings!
Shared Memory:

The shared memory regions are mapped via a ShMem struct inside the
ShMemCfg struct. The ShMem hereby behaves similar to a rust Vec. It
holds a ```*mut u8``` raw pointer to the memory area inside the PCI devices memory
area and a length field, indicating the length of the shared memory
area.
The ShMem struct implemnts Deref and DerefMut, which will return a
mutable or immutable &[u8]. This should allow to easily write/read
from/to the shared memory area.
Furthermore the struct implements a custom Drop, which will overwrite
the referenced memory area with zeros. (Hence, dropping the ShMem in a
different thread might be beneficial).

Fix of wrong mappings:

The mappings of the raw stuctures inside the devices memory space to its
structures here in the code where made like the following:

```Rust
raw_struct Box<RawStruct> = unsafe { Box::from_raw(*(address_of_struct as *mut
RawStruct))};
```

This resulted in a move of the raw pointer and in a copy to the
heap of the content lying at ```address_of_struct```. Which subsequently
was not what was indendet. Namely, allowing the manipulation of the data
inside the devices memory area via the strucutre.

Now the raw structures are referenced via a ```&'static mut```
reference. The reference can be seen as static as the strucutre which is
referenced belongs to the device and will be valid as long as the device
is present. Which we must/can assume.