Virtio network driver general init of device.
The Virtio network driver initalizes a virtio network device
according to the specification v1.1. - 3.1.1. This EXCLUDES the device
specific initalization of the network device, which is described in
the specification v1.1. - 5.1.5.

Currently the driver negotiates the following features with the device
(This features are also the minimal necessary set which MUST be provided
by a network device, otherwise it will be rejected)
- VIRTIO_NET_F_MAC
- VIRTIO_NET_F_STATUS
- VIRTIO_NET_F_GUEST_UFO
- VIRTIO_NET_F_GUEST_TSO4
- VIRTIO_NET_F_GUEST_TSO6
- VIRTIO_NET_F_GUEST_CSUM

This commit also includes a new virtio network driver specific error and
its wrapper in the VirtioError enum.
Also the ComCfg struct got some new functionality in order to have a
proper public interface for drivers. The interface will most likely be further
developed in the process of development of the driver.

Furthermore this commit cleans the Virtqueue section of the code, which
will be newly populated during the development of the network driver.