New dedicated FeatureSet struct for network driver.
Adds a new struct called FeatureSet to the NetDevCfg struct which
holds all features defined in the standard in Virtio specficication
v1.1. - 5.1.3.

The struct provides four functions:
- new(): Returns an new instance with all features set to false.
- set_feature(feat: Features) -> Result<(), VirtioNetError::FeatReqNotMet:
Sets the feature indicated by the enum
value of feat. In case the setting was not in set in accordance to
specification v1.1. - 5.1.3.1 the function will return an error.
- get_features() -> Option<Vec<Features>>: Returns a vector containing
all set features in form of their enum value.
- is_feature(feat: Features) -> bool: checks if a given feature is set.