Changed FeatureSet functionality.
This commit allows for FeatureSet to check a provided vector if
features. Hereby the function `check_features(&self, Vec<Features>)`
checks if the given features adhere to the specification v1.1. -
5.1.3.1. If they don't it returns an error.

As a result of the added function the old `set_feature(&mut self,
Features)` function was changed into `set_features(&mut self,
Vec<Features>)` which now omits a check and does allow to add all
Faetures in one step. Also the Features inside the vector does NOT need
to be in the "RiGHT" order anymore.