Added general feature bits to VirtioNetFeatures
In order to allow a complete check and setting of Features, this commit
adds the general Virtio features defined in spec v1.1. - 6 into the
Features enum of the virtio network features.
Those features also exist in the drivers/virtio/mod.rs file and as a
result introduce some duplication. Which is necessary to avoid dynamic
dispatch, otherwise needed to unify the functions of FeatureSet
for the two different enums in mod.rs und virtio_net.rs.