Public virtqueue interface for users and internals
The Virtq enum which provides a unified and public interface for drivers
to use either a packed or a split virqueue. The layout for this
interface is now finished, but might be changed or adjusted, during the
implemenatiton of the interface in the actual queues (Probably just some
getters and setters for the actual queues).

The implementation misses some parts, which are marked
`unimplemented()`. Theses parts will be added in future commits. Those
parts are left out as they refer to unsafe code, whichs form will be
clear in the future.

The individual parts of are documented extensively and will only be
breifly explained in the following:

The virtq enum does provide three main functions for drivers. The
functions allow to easly provide memory the driver. In case the driver
is ready to provide the memory to a device, the driver is able to
transform these memory areas into an ongoing transfer. This transfer
then can be polled and signals its state via an enum. Upon finish of the
transfer the state will indicate it and allow the driver to further
process the underlying memory areas.