Packed VQ: Prepare_{buffer,transfer,transfer_raw}
The packed virtqueues functions,
- prep_buffer()
- prep_transfer()
- prep_transfer_from_raw()

are finished.

This included functionality on MemDescr as well
as functionality on the MemPool. Which are now also allow to create
descriptors for already existing memory.

Furthermore some Virtqerrors were added and some docs.

INFO: Currently the implementation has some lifetime problems, but this
commit should safe the intermediate step. Lifetime problems will be
resolved in the next commit.