Fixed lifetime problems.
Beforehand each MemDescr carried a reference to the MemPool it comes
from. This resulted in a permanent borrow of the packed virtqueue
structure which holds the pool.
In order to prevent this scenario the MemDesc know contains a
Rc<MemPool> as does the virtqueue itself.
Furthermore in order to allow interior mutablity of the MemPool the pool
wraps its Vector of MemDescrId's into an RefCell.

This allows to call MemPool and borrow the vector mutable inside the
associated function.

In the future this patern should be exchanged with an Arc<MemPool> and
and Mutex<Vec<MemDescrId>> to allow for safe multithreading, as the
descriptor ID's are returned upon drop of a BufferToken, which can
possibly happen in multiple threads.