Fixed lifetime for generic in function body
The function body of the virtq interface function
`prep_transfer_from_raw` did not contain the needed static lifetime for
the gernerics which forces the user to give ownership to this function.

Furthermore this commit adds an empty function body to the Transfer
struct. This function will allow drivers to expect finished transfers at
a VecDequeue, this should make searching for finished transfers via
iteration through a list obsolete.