Exchanged &Virtq for Rc<Virtq> in BufferToken
Giving a BufferToken a refernce to the respective Virtq it is coming
from is essential for the functionality of the Queue in its current
design. But the refrences made it impossible to work with the queue
after the BufferToken were handed out to the driver.
Instead now, the Tokens hold Rc<Virtq> to the Virtq of the Driver.
This forces the driver to hold a Rc<Virtq> of the Virtq himself.
This changed forced changing all generic lifetime parameters in the
commited files.

Furthermore this commit adds functionality for different structures of
the Virtq and the PackedVq implementation. Like:

* dispatch_await() for TransferTokens
	-> This allows TransferTokens to be send and awaited at VecDeque
* returning data from Finished Transfers.
* dispatching TransferTokens in general.

Also fixes a logical error in the creation of MemDescr, which used a
wrong casting, which resultet in unusable raw ptrs.