Changed enum Buffer tp c-like structs.
This changes the enum Buffer which holded three types:
* Single,
* Multiple,
* Indirect,

The enum still contains the same variants allthough they are now, not
holding a list of memory descriptors but instead a list of memory
descriptors, a len field, indicating the overall bytes in the buffer and
a last_write field in order to allow write sequentially into a buffer.

This change also eases the computation of the length of a buffer, which
otherwise would result in unnecessary and repeated iteration through the
list of memory descriptors inside a buffer.