BufferToken are writable now.
The underlying buffers of a BufferToken are now writable via
* write
* write_seq

Both functions will fail if the write to the buffer was to large and
return an VirtqError. For details see docs.

In order to facilitate this easily this commit adds some convenicnce
functionality to the Buffers struct. As also an own error type for
Buffers.