Utility functions for Phy Page Boundary Checks
This commit adds two utility functions to the Virtq enum.
The functions allow to check wheter a given structure or a given slice
do cross a non-contigious physical page boundary, which is not permitted
for data structures provided to the queue. The queue will panic if this
happens.

Drivers should use these functions in order to ensure, that this does
not happen.

In the future there might be an option, where the queue will copy
structures, violating these conditions instead of panic-ing.