VirtioDriver enum for lifetime issues
The compiler was not able to correctly infer, that the code enum
PciDriver does not contain the lifetime of the PciAdapter struct that is
used as an argument during virtio device initalization. As the register
driver function has a 'static generic lifetime parameter.
In order to register drivers in the kernels pci code, Virtio driver
structs are now wrapped inside an VirtioDriver enum. This enum is
returned upon initalization and then outside of the init function, where
the lifetime of the &PciAdapter was elided into, the Virtio drivers are
destructured and Wrapped inside the PciDriver enum and provided via the
register function to the kernel.