Control General Notifications
This commit provides the packed virtqueue with the functionallity to
disable and enable general notifications by the device, which the device
sends when it has marked an buffer as used.

Furthermore it provides functionality to enable a notification only or
also for a specific transfer. This option is only valid if
VIRTIO_F_EVENT_IDX has been negotiated, also the queue will not fail if
one signals a specific notification is desired and the feature has not
been negotiated.
This is due to the fact, that one can not relly on this notifications to
happen 100%.

Also in this commit:
- possibility to enable and disable queues selecively.
  - Queues are enabled by default upon creation.
- adjustments inside the virtio_net driver, due to the changes of the queue.