Change allocation method for MemDescr inside Pool
This commit changes the method how memory is allocated for MemDescr
inside the MemPool.

In order to ensure contignous memory the method uses the
crate::mm::allocate() function.

Furthermore this commit changes the Deallocation of MemDescr as it is
important to deallocate memory via either crate::mm::deallocate() or
simply via Vec::from_raw_parts(). Depending on whether the memory came
from a structure alreadyd allocated or from freshly allocated memory.

Also this commit fixes an eroneous function for Transfers which allowed
to directly retrieve the access to the memory areas of a buffer.
This function now returns a Vec<Vec<u8>> in order to prevent
deallocation errors.
As this functionality is highly experimentially the function is marked
unsafe and the docs further indicate the instability of this call!