Changed Handling to main driver area
This commit moves responisbility to the correct area. Namely the main
network driver now handles how buffers are handled upon reception and
transmission. The RxQueues and TxQueues are only responisble for
providing the next Tokens/Transfer

Additionally this commit adds the first structure to provide the driver
with control functionalitChanged Handling to main driver area

This commit moves responisbility to the correct area. Namely the main
network driver now handles how buffers are handled upon reception and
transmission. The RxQueues and TxQueues are only responisble for
providing the next Tokens/Transfer

Additionally this commit adds the first structure to provide the driver
with control functionalityy