Removed unnecessary copy network driver
This commit allows to remove an unnecessary copy of data between the
kernel-space driver and the user-space smoltcp stack.
In order to allows this Transfers are now capable of returning a vector
which hold refernces to the memory areas as u8 slices.

Also this commit changes the rx-buffers. Thei are now two descriptors
long. One in the size of the virtio header and one in the size of the
complete packet size exepcted by the device. THis is due to the fact,
that we currently do not change this sice during runtime and only need
to set up (i.e. allocate) buffers at initalization if the driver.