Logical Bug fixes in Driver, Virtqueue and Env
- Fixes bug in environment, where failure occured when usize was created
from a u32 on a u64 system. Now all calls are made via usize::try_from()
and are directly unwrapped to ensure quick failure on non supported
platforms.
- Fixes bug in network driver, where during feature negotiaten the
driver did reduce its features not with the devices features but with
himself, which of course did not reduce the feature set at all.

Fixes bug in the virtqueue:
- raw_ptrs function of BufferToken, which did not returne the correct
data due to a wrongly match statement. Always returned None, where Some
was expected. And Some(empty_vector), where None was expected.
- Changes MemDesc to hold its inital descriptor length and to correctly
hold its length. The first is needed in order to correctly reuse a
buffer, as the device might change the length field after writing to a
buffer. The second did hold the length of the allocated memory area,
which might differ from the wanted length (allocated memory area is
always larger or equal to wanted length) due to the underlying
allocation function of the kernel.

Furthermore the documentation was extended in order to reflect these
changes.