Bug Fixes
VirtioNet Driver
- logic error in for loop of virtqueue init, where the driver should
create vqs in pairs of even and uneven indexes. This was wrong and did
not start at 0 but at 1.
- From the above bug another bug resided in the creation of the control
vq, which used an index: 2*num_vqs+1. This should have been 2*num_vqs.

Pci:
- During the read of the 64 bit field, which indicates the devices
features, the pci code missed a shift of the higher 32 bits to the
actual higher 32 bits position inside the 64 bit field

Virtqueue:
- AsSliceU8 trait had a faulty default implementation, where the
computation of the actual size of the given struct, the implementation
calculated the sive of the reference (i.e. the size of a usize).
- BufferToken.write_seq() method needed and explicit deref() call upon
the MemDescr in order to correctly execute a copy_from_slice() call.