Bug Fixes
This is another commit in a series of BugFixes.
The fixes and changes are detailed in the section below.

virtio_net
- Changed receive_rx_buffer to take into accoutn the possibility of a
zero sized buffer, which will result in an error return.

virtio_mod:
- Added functionaity to restrict the Buffers size. This size is actual
not important when it comes to accesing the actual memory areas but
might be used by drivers for checks. Hence it is important to have a
consistent state here, as the ReadCtrl inside the packed virtuqeue will
change the accesible memory area if devices write smaller amounts of
data into the buffers than the buffers size.
- Added functionality to restore the Buffers size at initalization time.
This is used when BufferTokens are reused via Transfer.reuse()...

Pci:
- Small bug in pci code, were I did modify a bit wrongly.

PackedVQ:
- Changed the working of ReadCtrl to not rely on the queue but instead
of the given BufferTokens to check if a transfer used indirect
descriptors and so on. Changed the ReadCtrl furthermore to restrict the
BufferTokens Buffers sizes (actually just receive buffers, as others are
read only by the queue) to the size writen by the device.
- Furthermore the ReadCtrl currently does bring the DescriptorRing into
a state where all data is deleted after a transfer, besides the right
Avail/used flags.
    -> Testing needs to be made in order to check, if any devices do
    rely on this data and then, we should fall back to only updating the
    flags, or not even that.