Changed logic error in poll function
The old poll function did created a Pinned<TransferToken> from the *mut
TransferToken reference stored in the Descriptor ring. Then it set the
state to finished and decided if we either provide the token into its
await_queue or forget it.
This procedure had one major problem:

- leaked memory as a vector of Pinned<TransferToken> was simply leaked.
- Was undefined behaviour, as it can not be ensured, that a drier would
not directly drop the Transfer, which would result in a drop of the
TransferToken too, upon receiving the state of the transfer changed to
finished.

Now we do not create a Pinned<TransferToken> up to the point, where we
can ensure, that the driver does not hold a Transfer of the respective
TransferToken but awaits the Transfer at the await_queue. Furthermore we
just operate on the reference up to the point, where we set the state to
finished.