Allow Pinned<T> into *mut T
This commit allows pinned T to be converted into raw pointers. This is
mostly needed in order to prevent the Pinned<T> to keep ownership of T
and enforece deallocation upon drop. Hence it can be used to prevent
deallocation of the pinned type.

This is currently needed in await transfers. As those functions call the
normal dispatch function of the virtqueues which return a transfer, we
can now simply take the pinned<T> from the transfers and convert them
into raw pointers and drop them. This is memory safe and does not leak
any memory.