Fixed Bug in Virtio Notification Setup
The bug resulted from the fact, that notifications are enabled by
setting a zero and are disabled by setting a 1. This is fixed now.
Furthermore the reading of the notification multiplier in the virtio pci
code was eronous. A typo resulted in a read of the wrong 32 bit value.
Multiplier is now read correctly. Also the creation of the notification
data was wrong, as it placed the bits at the wrong position. THis is now
fixed.