Vnet Driver: Feature adjustment, receive of data
This commit changes a few small thing:

The network driver is currently not able to support
VIRTIO_NET_F_GUEST_CSUM and all features depending in this one.
it is deactivated here. In order to be able to enable this feature in
the future I added a post_process function to the RxQueues, which will
take care of calculating checksums and could possibly merge buffers to.
Currently it is simply doing nothing.

Furthermore I sligtly adjusted the virtio pci code in oder to only
register an interrupt handler for the network driver, when the network
driver has been created succesfully.

Additionally due to the behaviour of the network device and the
implemenattion in the linux kernel I ommitted the check of the WRITE
flag in used descriptors and simply assume that, when length has been
set in used descriptors, the buffer has been written.