Fixed bug in add RxQueues
In case where segmentation offload is not used, the code used the wrong
buffer specification and a multiple buffer. This is now changed to a
indirect one or multuple.

Furthermore during getting of a buffer, we allow now the resizing of
tokens and also get a single buffer in case nothing fits and not a
multiple one.

Aslso some changes in the docs.