Changed time when poll queue network driver get_tx
THis commit only polls the queue, during get_tx_buffer when the await
queue is empty.
One could assume now, that we poll the queue everytime, the complete
queue is once available and then used for the device. Then the driver,
will initalize a poll, which results in a read of all used descriptors
, which are then placed into the drivers awaiting structure.