Finished Contigunous Memory for PackedVq
This commit finishes the integration of the new memory layout for the
packed virtqueue.
In the past, the queu used to allocate its memory via a vector. As this
memory is not gueranteed to be contignuous in physical address space,
this commit changes the memory allocation to be allocated by the kernel,
which ensures continuity.