network drivers create directly TxToken / RxToken for smoltcp
For this purpose, it was necessary to avoid the creation of dynamic objects of the driver. This means that during compilation it must be defined which network interface is used. If the feature `rtl8139` is activated, the driver for Realtek 8139 will be used. Otherwise the Virtio network interface will be used.

using a exponential backoff to check the network device

if exponential backoff has completed, the task will be blocked
and waits for an network interrupt

poll device before the device checks the future

remove obsolete code

network drivers create directly TxToken / RxToken for smoltcp

For this purpose, it was necessary to avoid the creation of dynamic
objects of the driver. This means that during compilation it must
be defined which network interface is used. If the feature `rtl8139`
is activated, the driver for Realtek 8139 will be used. Otherwise the
Virtio network interface will be used.

build Virtio network driver only if the feature `tcp` is set

feature `tcp` must be activated to support any network interface

remove compiler warnings

remove compiler warnings

remove compiler warnings

use polling mode for read / write operations on a socket

idle task enables network interrupt, while the read function disable
it

enable network interrupts, if a socket is blocking

add missing import

check device driver only, if the device is available

remove compiler warnings

add mmio support

add dummy network interface for aarch64

add missing file for mmio support

add check if the feature tcp is available

revert block_on to the original version

revise CI workflow to support the feature rtl8139