perf(alloc): don't pad the size of allocations
Increasing the alignment of an allocation is sufficient to avoid false sharing.
Talc needs a little bit of metadata next to each allocation, and this would create N-8 sized holes in between every allocation due to the demands placed on it.
By removing the size requirement, false sharing avoidance is unhindered, but this allows Talc to minimize the overhead of high-alignments.
(Talc is designed to automatically place the metadata in a way that minimized false sharing, but this only helps if Talc can sneak it in within the same cache line, which is impossible if Talc thinks you want the whole cache line for user data.)

Co-authored-by: Martin Kröning <martin.kroening@eonerc.rwth-aachen.de>
Signed-off-by: Martin Kröning <martin.kroening@eonerc.rwth-aachen.de>