virtq: remove TransferState
The TransferState was needed to check the status of the Transfer when it was shared between the caller and the callee. However, this is no longer the case and the caller cannot access a in-progress transfer anymore. Thus, the related checks are no longer needed and the field is no longer necessary.