virtq: delay MemDescrId assignment until push
Not every MemDescr necessarily has its own id in the queue. For example, descriptors that make up a chain in a packed queue share their Buffer ID. Descriptors in an indirect table have ids but they don't take them from their queue's pool. Not making the id a field of the MemDescr also removes the coupling between the MemDescr and a queue.