virtqueue/packed: don't modify flags for skipped descriptors
`reset_ring_pos` has an effect only when a used descriptor's flag has not been already marked as used by the device (i.e. a used descriptor for a chained buffer had been written and the descriptor in question was skipped). Since these descriptors are skipped by the read controller and the write controller relies on the capacity field for descriptor slot availability (rather than the flags), it is not necessary to do this modification. There also does not seems to be anything in the specification to indicate this modification is necessary.