refactor(virtq): use safe types for virtq buffers
Instead of treating any object that is transmitted through virtqueues as
a byte slice, use `Box<dyn Any>` when possible. This removes the need to
handle allocations and deallocations inside the Virtq module and rely on
what is provided by Rust. It also allows us to handle the elements in a
more type-safe manner.