refactor(virtq): represent avail and used buffers as distinct types
Sized element downcasting and vector popping from the device-writable buffer only make sense for buffers that have been used by the device. By separating the available buffer and the used buffer types, we can ensure that they are called at the correct point of the buffer's lifecycle and make the assume_init calls for the driver. Furthermore, a distinct used buffer type allows us to delay the initialized length checks until the point of use by the driver, where the potential errors can be better handled.