add support of contiguous bit
The arm64 MMU supports a contiguous bit which is a hint that the TTE
is one of a set of contiguous entries which can be cached in a single
TLB entry.

To support this feature, the function map checks if cntinguous region
has to be mapped into the address space.
In addition, the function map_heap tries to allocate large memory
blocks to increase the propobility of contiguos memory regions.