perf(virtio-net): Don't copy packets when num_buffers is 1
Currently, we would copy every single received packet from DeviceAlloc
to a new Vec that is allocated with GlobalAlloc. If we use DeviceAlloc
for the payload of the RxToken, we can avoid copying all packets for the
special (and common) case of num_buffers = 1.

While at it, I refactored the code to make it more readable and avoid
allocating the intermediate Vec that would be collected into a flattened
Vec<u8> anyways.

Signed-off-by: Jens Reidel <adrian@travitia.xyz>